On branch main 	modified:   index.html 	modified:   main.js 	modified:   node_modules/.package-lock.json 	modified:   node_modules/dns-packet/index.js 	modified:   node_modules/dns-packet/package.json 	modified:   node_modules/ws/lib/websocket-server.js 	modified:   node_modules/ws/package.json 	modified:   package-lock.json 	modified:   src/audio_core/audio_base.js 	modified:   src/audio_core/audio_inputs.js 	new file:   src/audio_core/export_wav.js 	new file:   src/audio_core/metronome.js 	deleted:    src/audio_core/play_slice.js 	new file:   src/audio_core/sample_player.js 	modified:   src/audio_core/sync_tempo.js 	deleted:    src/audio_core/track_watchdog.js 	modified:   src/create_track.js 	modified:   src/elements/basic_elements.js 	modified:   src/elements/create_slicers.js 	deleted:    src/elements/create_table.js 	modified:   src/elements/create_trim_handlers.js 	new file:   src/elements/metronome_box.js 	modified:   src/elements/record_box.js 	modified:   src/idb.js 	modified:   src/index.js 	modified:   src/keyboard_events.js 	modified:   src/populate_nav_menu.js 	deleted:    src/record_functions.js 	new file:   src/style/_library.scss 	new file:   src/style/_metronome_box.scss 	new file:   src/style/_nav_menu.scss 	renamed:    src/style/record_window.scss -> src/style/_record_box.scss 	renamed:    src/style/track.scss -> src/style/_track.scss 	deleted:    src/style/nav_menu.scss 	modified:   src/style/style.scss

+++ src/populate_nav_menu.js
@@ -2,23 +2,22 @@ import {createNumberBox,
         createButton,
         createSpan,
         createSlider,
+        createToggle,
         sliderRuleMarks} from './elements/basic_elements.js';
-import createTable from './elements/create_table.js';
 import createRecordBox from './elements/record_box.js';
-import {setBPM} from './audio_core/sync_tempo.js';
+import createMetronomeBox from './elements/metronome_box.js';
 import {setMasterVol} from './audio_core/audio_base.js';
 import {deleteTrack} from './create_track.js';
-
 import {stopRecStream,
         startRecStream} from './audio_core/audio_inputs.js';
 
 const menu = document.querySelector('.nav_menu');
 menu.appendChild(createButton('Record Now',['open_record_box'],openRecording));
-menu.appendChild(createNumberBox('BPM:','set_bpm',[30,600],setBPM,120));
-const master_vol = createSlider('Master: ','master_volume',[0,1],0.005,setMasterVol,1);
-menu.appendChild(master_vol);
-menu.appendChild(createSpan('0.00dB','decibels'));
 menu.appendChild(createButton('Delete All',['delete_all'],deleteAllTracks));
+const master_vol = createSlider('Master: ','master_volume',[0,1],0.005,setMasterVol,1);
+master_vol.appendChild(createSpan('0.00dB','decibels'));
+menu.appendChild(master_vol);
+menu.appendChild(createMetronomeBox());
 menu.appendChild(createRecordBox());
 
 function openRecording() {
@@ -38,6 +37,8 @@ function openRecording() {
 }
 
 function deleteAllTracks() {
-    const all = document.querySelectorAll('.track');
-    all.forEach(track => deleteTrack(track))
+    if (confirm('Are you sure?')) {
+        const all = document.querySelectorAll('.track');
+        all.forEach(track => deleteTrack(track));
+    }
 }
